Added Start of HTML Pages And Controller
Added all the HTML pages I could immediately think of for the site. Not all will be shown at any given time as we can use the "th:if" conditional attribute to make it so only some of the links are shown based on whether a user is loggin in or not. Controller for serving the pages is setup and pages have releveant text but are otherwise unpopulated.