Added Front And Back End Connectivity
Added basic connectivity between the front and back end for account creation and logging in. Created forms on the front end which post to the relevant controllers in the back. Added ways of verifying if a user with a username exists or not and checks if their password is correct. Also added password hashing when creating accounts.