Added Error Message For Incorrect User Details
Added a generic error message without too much info to be displayed if a username and password do not match an existing entry in the database. Used flashAttributes instead of model.addAttribute because these will only be there for the one redirect and will be removed once successfully logged in.