Added Documentation For User Login Method
Added comments for what the method, class and parameters do for the login post handler. Verified that user is added to model after being retrieved. Verified that created accounts are retrieved and logged into. Closing user login related issues as all seem to be completed with this commit. Closes #2, closes #3, closes #4.