(Hopefully) Fixed Errors Loading Pages
Due to flat lazy loading there were a couple errors that stopped thymeleaf from working and as such pages wouldn't load. I've added custom queries to use in the place of user.getFlat() that also retrieve all the things joined to a flat like the noticeboard and flatMemberships. Use them or create one similar if you run into this issue.